Provided framework for the timing classes and check analysis classes
I added the framework for all three timing classes and check analysis classes. As of right now all that code is copied directly from the class examples and will need to be changed accordingly. I also mentioned in the JavaDoc comments for each of these classes that the finished product will be modified code from lecures 6 and 7 as to cite our sources.