fixed bugs with sort() and insertionSort()
There were several bugs with sort and insertion sort mehods. Switched sort so that it now uses String.split("") to create an array of String objects to represent the individual characters of the string. This is better and more concise than what it was before and allows to use toLower() to fix bugs with uppercase letters.
Another major bug was that the initial currentIndex variable was starting at 1 instead of 0 in the insertion sort algorithm. These are fixed now.
There are still many tests that fail in the JUnit testing but this is mostly because of problems with the tests themselves.
I fixed the easy problems with JUnit tests, but most of them need to be reimplemented, and the ones that require text files need the text files to be created for them.